fix: split loom-daemon into lib + binary to prevent test hang
The loom-daemon test binary hung on startup (even `--list`) because tests
were compiled into the binary target alongside the tokio runtime and IPC
server setup. Moving all modules into a library crate ensures tests run
against the library target, which has no main() function or blocking
initialization.

- Create src/lib.rs with all module declarations and utility functions
- Slim src/main.rs to just the binary entry point (no #[cfg(test)])
- Add crate-level clippy allows for pre-existing pedantic lints newly
  exposed by the library's public API surface

Closes #2283
