+++ loom-daemon/src/terminal.rs
@@ -232,6 +232,7 @@ mod claude_config {
     }
 
     #[cfg(test)]
+    #[allow(clippy::unwrap_used)]
     mod tests {
         use super::*;
